refactor: Add "token" field to UserResource toArray method
The code changes in the UserResource.php file add a "token" field to the toArray method. This allows the "token" attribute to be included in the JSON response when the UserResource is used to transform user data. This enhancement improves the UserResource by including the "token" attribute in the response when it is not null.

Note: This commit message follows the established convention of using a prefix (refactor) to indicate code refactoring or improvement.